+++ src/com/PluginBase/ParticleEmitter.java
@@ -4,6 +4,7 @@ import org.bukkit.Location;
 import org.bukkit.Particle;
 import org.bukkit.Particle.DustOptions;
 import org.bukkit.World;
+import org.bukkit.block.data.BlockData;
 import org.bukkit.entity.Entity;
 import org.bukkit.plugin.Plugin;
 import org.bukkit.scheduler.BukkitRunnable;
@@ -209,7 +210,12 @@ public class ParticleEmitter {
 			if (particle == particleThatNeedsData) {
 				
 				// Log the error in the console
-				Chat.getInstance().sendErrorMessageToConsole(null, "Tried to emit particle that needs data: Particle." + particle + " may need additional dust options");
+				Chat.getInstance().sendErrorMessageToConsole
+				(
+						null,
+						"Tried to emit particle that needs data:"
+						+ "Particle." + particle + " may need either dust options or block data to be spawned"
+				);
 				
 				// Do not actually spawn the particle
 				return;
@@ -231,7 +237,7 @@ public class ParticleEmitter {
 	}
 	
 	/**
-	 * Emit particles once at a specific location
+	 * Emit redstone dust once at a specific location
 	 * 
 	 * @param world 		The world in which the particles will be emitted
 	 * @param location 		The location at which the particles will be emitted
@@ -239,7 +245,7 @@ public class ParticleEmitter {
 	 * @param amount 		How many particles should be emitted
 	 * @param speed 		How quickly the particles should play their animation
 	 * @param spread 		The spread of the particles (using the entity location as a base)
-	 * @param type			The type of the particle
+	 * @param dustOptions	The type of the particle dust
 	 */
 	public void emitParticles(World world, Location location, Particle particle, DustOptions dustOptions, int amount, double speed, Vector spread) {
 		world.spawnParticle
@@ -253,7 +259,34 @@ public class ParticleEmitter {
 				spread.getY(),				// Particle spread Y
 				spread.getZ(),				// Particle spread Z
 				speed,						// Particle speed
-				dustOptions					// Particle type
+				dustOptions					// Particle dust options
+		);
+	}
+	
+	/**
+	 * Emit particles with custom block data once at a specific location
+	 * 
+	 * @param world 		The world in which the particles will be emitted
+	 * @param location 		The location at which the particles will be emitted
+	 * @param particle 		The particle type to be emitted
+	 * @param amount 		How many particles should be emitted
+	 * @param speed 		How quickly the particles should play their animation
+	 * @param spread 		The spread of the particles (using the entity location as a base)
+	 * @param blockData		The type of block the particles should belong to
+	 */
+	public void emitParticles(World world, Location location, Particle particle, BlockData blockData, int amount, double speed, Vector spread) {
+		world.spawnParticle
+		(
+				particle,					// Particle type
+				location.getX(),			// Location X
+				location.getY(),			// Location Y
+				location.getZ(),			// Location Z
+				amount,						// Particle amount
+				spread.getX(),				// Particle spread X
+				spread.getY(),				// Particle spread Y
+				spread.getZ(),				// Particle spread Z
+				speed,						// Particle speed
+				blockData					// Particle block data
 		);
 	}
 }